+++ rollup.config.js
@@ -1,14 +1,17 @@
-import commonjs from 'rollup-plugin-commonjs'
 import buble from 'rollup-plugin-buble'
 import { uglify } from 'rollup-plugin-uglify'
 
 export default {
   input: './src/index.js',
   plugins: [
-    commonjs(),
     buble(),
     uglify()
   ],
+  watch: {
+    chokidar: true,
+    include: 'src/**',
+    clearScreen: false
+  },
   output: {
     format: 'umd',
     name: 'vueDebounce',